+++ v3/src/assets/icons/github-octicons/README.md
@@ -1,0 +1,1 @@
+https://github.com/primer/octicons
