+++ v3/src/components/graph/components/inspector-panel/graph-measure-group.scss
@@ -10,13 +10,13 @@
 
 .measure-items {
   margin-left: 20px;
-}
 
-.measure-movable-value-button {
-  font-size: 12px!important;
-  font-weight: 400!important;
-  background-color: #eeeeee!important;
-  height: 20px!important;
-  width: fit-content;
-  margin-bottom: 5px;
+  .measure-movable-value-button {
+    font-size: 12px!important;
+    font-weight: 400!important;
+    background-color: #eeeeee!important;
+    height: 20px!important;
+    width: fit-content;
+    margin-bottom: 5px;
+  }
 }
