+++ v3/src/models/ui-state-types.ts
@@ -1,0 +1,4 @@
+export const RulerStateKeys = ["measuresOfCenter", "measuresOfSpread", "boxPlotAndNormalCurve", "otherValues"] as const
+export type RulerStateKey = typeof RulerStateKeys[number]
+
+export type RulerState = Record<RulerStateKey, boolean>
